Refactor: RecipeProcess 리팩토링
