+++ src/main/java/org/springeel/oneclickrecipe/domain/recipeprocess/dto/controller/RecipeProcessUpdateControllerRequestDto.java
@@ -3,7 +3,6 @@ package org.springeel.oneclickrecipe.domain.recipeprocess.dto.controller;
 public record RecipeProcessUpdateControllerRequestDto(
     Byte sequence,
     String description,
-    String imageUrl,
     Short time
 ) {
