+++ src/main/java/org/springeel/oneclickrecipe/domain/recipeprocess/dto/service/request/RecipeProcessCreateServiceRequestDto.java
@@ -1,4 +1,4 @@
-package org.springeel.oneclickrecipe.domain.recipeprocess.dto.service;
+package org.springeel.oneclickrecipe.domain.recipeprocess.dto.service.request;
 
 public record RecipeProcessCreateServiceRequestDto(
     Byte sequence,
