+++ src/main/java/org/springeel/oneclickrecipe/domain/recipeprocess/dto/service/request/RecipeProcessUpdateServiceRequestDto.java
@@ -1,9 +1,8 @@
-package org.springeel.oneclickrecipe.domain.recipeprocess.dto.service;
+package org.springeel.oneclickrecipe.domain.recipeprocess.dto.service.request;
 
 public record RecipeProcessUpdateServiceRequestDto(
     Byte sequence,
     String description,
-    String imageUrl,
     Short time
 ) {
