+++ src/main/java/org/springeel/oneclickrecipe/domain/recipeprocess/dto/service/response/RecipeProcessReadResponseDto.java
@@ -1,4 +1,4 @@
-package org.springeel.oneclickrecipe.domain.recipeprocess.dto.service;
+package org.springeel.oneclickrecipe.domain.recipeprocess.dto.service.response;
 
 public record RecipeProcessReadResponseDto(
     Byte sequence,
