+++ src/main/java/org/springeel/oneclickrecipe/domain/recipeprocess/mapper/dto/RecipeProcessDtoMapper.java
@@ -1,13 +1,15 @@
 package org.springeel.oneclickrecipe.domain.recipeprocess.mapper.dto;
 
+import static org.mapstruct.MappingConstants.ComponentModel.SPRING;
+
 import org.mapstruct.Mapper;
 import org.mapstruct.MappingConstants.ComponentModel;
 import org.springeel.oneclickrecipe.domain.recipeprocess.dto.controller.RecipeProcessCreateControllerRequestDto;
 import org.springeel.oneclickrecipe.domain.recipeprocess.dto.controller.RecipeProcessUpdateControllerRequestDto;
-import org.springeel.oneclickrecipe.domain.recipeprocess.dto.service.RecipeProcessCreateServiceRequestDto;
-import org.springeel.oneclickrecipe.domain.recipeprocess.dto.service.RecipeProcessUpdateServiceRequestDto;
+import org.springeel.oneclickrecipe.domain.recipeprocess.dto.service.request.RecipeProcessCreateServiceRequestDto;
+import org.springeel.oneclickrecipe.domain.recipeprocess.dto.service.request.RecipeProcessUpdateServiceRequestDto;
 
-@Mapper(componentModel = ComponentModel.SPRING)
+@Mapper(componentModel = SPRING)
 public interface RecipeProcessDtoMapper {
 
     RecipeProcessCreateServiceRequestDto toRecipeProcessCreateServiceRequestDto(
