+++ src/main/java/org/springeel/oneclickrecipe/domain/recipeprocess/service/RecipeProcessService.java
@@ -2,9 +2,9 @@ package org.springeel.oneclickrecipe.domain.recipeprocess.service;
 
 import java.io.IOException;
 import java.util.List;
-import org.springeel.oneclickrecipe.domain.recipeprocess.dto.service.RecipeProcessCreateServiceRequestDto;
-import org.springeel.oneclickrecipe.domain.recipeprocess.dto.service.RecipeProcessReadResponseDto;
-import org.springeel.oneclickrecipe.domain.recipeprocess.dto.service.RecipeProcessUpdateServiceRequestDto;
+import org.springeel.oneclickrecipe.domain.recipeprocess.dto.service.request.RecipeProcessCreateServiceRequestDto;
+import org.springeel.oneclickrecipe.domain.recipeprocess.dto.service.response.RecipeProcessReadResponseDto;
+import org.springeel.oneclickrecipe.domain.recipeprocess.dto.service.request.RecipeProcessUpdateServiceRequestDto;
 import org.springeel.oneclickrecipe.domain.user.entity.User;
 import org.springframework.web.multipart.MultipartFile;
