+++ src/main/java/org/springeel/oneclickrecipe/domain/recipeprocess/service/impl/RecipeProcessServiceImpl.java
@@ -8,9 +8,9 @@ import org.springeel.oneclickrecipe.domain.recipe.entity.Recipe;
 import org.springeel.oneclickrecipe.domain.recipe.exception.NotFoundRecipeException;
 import org.springeel.oneclickrecipe.domain.recipe.exception.RecipeErrorCode;
 import org.springeel.oneclickrecipe.domain.recipe.repository.RecipeRepository;
-import org.springeel.oneclickrecipe.domain.recipeprocess.dto.service.RecipeProcessCreateServiceRequestDto;
-import org.springeel.oneclickrecipe.domain.recipeprocess.dto.service.RecipeProcessReadResponseDto;
-import org.springeel.oneclickrecipe.domain.recipeprocess.dto.service.RecipeProcessUpdateServiceRequestDto;
+import org.springeel.oneclickrecipe.domain.recipeprocess.dto.service.request.RecipeProcessCreateServiceRequestDto;
+import org.springeel.oneclickrecipe.domain.recipeprocess.dto.service.request.RecipeProcessUpdateServiceRequestDto;
+import org.springeel.oneclickrecipe.domain.recipeprocess.dto.service.response.RecipeProcessReadResponseDto;
 import org.springeel.oneclickrecipe.domain.recipeprocess.entity.RecipeProcess;
 import org.springeel.oneclickrecipe.domain.recipeprocess.exception.NotFoundRecipeProcessException;
 import org.springeel.oneclickrecipe.domain.recipeprocess.exception.RecipeProcessErrorCode;
@@ -38,14 +38,14 @@ public class RecipeProcessServiceImpl implements RecipeProcessService {
     @Value("${cloud.aws.s3.bucket.name}")
     public String bucket;
 
-
+    @Override
     public void createRecipeProcess(
         final RecipeProcessCreateServiceRequestDto requestDto,
         User user,
         Long recipeId,
         MultipartFile multipartFile
     ) throws IOException {
-        Boolean validator = recipeProcessRepository.existsBySequence(requestDto.sequence());
+        boolean validator = recipeProcessRepository.existsBySequence(requestDto.sequence());
         if (validator) {
             throw new ValidateRecipeProcessException(RecipeProcessErrorCode.USE_VALIDATE_DATA);
         }
@@ -68,6 +68,7 @@ public class RecipeProcessServiceImpl implements RecipeProcessService {
         s3Provider.saveFile(multipartFile, fileUrl);
     }
 
+    @Override
     public void deleteRecipeProcess(
         Long recipeId,
         User user,
@@ -84,6 +85,7 @@ public class RecipeProcessServiceImpl implements RecipeProcessService {
         s3Provider.deleteImage(recipe.getFolderName() + imageName);
     }
 
+    @Override
     @Transactional
     public void updateRecipeProcess(
         final RecipeProcessUpdateServiceRequestDto requestDto,
@@ -107,14 +109,13 @@ public class RecipeProcessServiceImpl implements RecipeProcessService {
         );
     }
 
+    @Override
     public List<RecipeProcessReadResponseDto> readRecipeProcess(
         Long recipeId) {
         Recipe recipe = recipeRepository.findById(recipeId)
             .orElseThrow(() -> new NotFoundRecipeException(RecipeErrorCode.NOT_FOUND_RECIPE));
         List<RecipeProcess> recipeProcesses = recipeProcessRepository.findAllByRecipeOrderBySequenceAsc(
             recipe);
-        List<RecipeProcessReadResponseDto> readRecipeProcess =
-            recipeProcessEntityMapper.toReadRecipeProcess(recipeProcesses);
-        return readRecipeProcess;
+        return recipeProcessEntityMapper.toRecipeProcessReadResponseDto(recipeProcesses);
     }
 }
